chore(jsnam) : 페이지, 서치 api 디렉토리 수정
