--- src/entities/community/api/index.ts
@@ -1,3 +1,0 @@
-export { fetchGetList } from './ListGet';
-export { useGetPageList } from './useGetPageList';
-export { useGetSearchList } from './useGetSearchList';
