+++ src/features/community/list/lib/index.ts
@@ -1,0 +1,2 @@
+export { useGetPageListQuery } from './useGetPageListQuery';
+export { useGetSearchListQuery } from './useGetSearchListQuery';
